requirements.txt
google-generativeai==0.4.0
pillow==11.1.0
streamlit==1.31.1
streamlit.option-menu==0.3.12






